+++ server/src/graql/internal/executor/QueryOperationExecutor.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.graql.internal.query.executor;
+package grakn.core.graql.internal.executor;
 
 import grakn.core.server.Transaction;
 import grakn.core.graql.concept.Concept;
