+++ server/src/graql/internal/match/MatchInfer.java
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.core.graql.internal.query.match;
+package grakn.core.graql.internal.match;
 
 import grakn.core.server.exception.GraqlQueryException;
 import grakn.core.graql.Match;
