+++ server/src/graql/internal/pattern/property/PropertyExecutor.java
@@ -21,7 +21,7 @@ package grakn.core.graql.internal.pattern.property;
 import grakn.core.graql.concept.Concept;
 import grakn.core.graql.Var;
 import grakn.core.graql.admin.VarProperty;
-import grakn.core.graql.internal.query.executor.QueryOperationExecutor;
+import grakn.core.graql.internal.executor.QueryOperationExecutor;
 import com.google.auto.value.AutoValue;
 import com.google.common.collect.ImmutableSet;
